# week8/workouv1/swagger.yaml
openapi: 3.0.0
info:
  title: Workouts API
  contact: {}
  version: 1.0
servers:
  - url: http://localhost:4000/api
    variables: {}
paths:
  /workouts:
    post:
      tags:
        - Workouts
      summary: Create Workout
      operationId: CreateWorkout
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
            example:
              title: Sample Workout
              reps: 14
              load: 50
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Workouts
      summary: Get All Workouts
      operationId: GetAllWorkouts
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  "/workouts/{id}":
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The Workout ID
    get:
      tags:
        - Workouts
      summary: Get Single Workout
      operationId: GetSingleWorkout
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Workouts
      summary: Update Workout
      operationId: UpdateWorkout
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkoutRequest"
            example:
              title: Random Workout
              reps: 1
              load: 5
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Workouts
      summary: Delete Workout
      operationId: DeleteWorkout
      parameters: []
      responses:
        "204":
          description: ""
          headers: {}
      deprecated: false
components:
  schemas:
    CreateWorkoutRequest:
      title: CreateWorkoutRequest
      required:
        - title
        - reps
        - load
      type: object
      properties:
        title:
          type: string
        reps:
          type: integer
        load:
          type: integer
      example:
        tille: Push ups
        reps: 40
        load: 5
    UpdateWorkoutRequest:
      title: UpdateWorkoutRequest
      required:
        - tille
        - reps
        - load
      type: object
      properties:
        tille:
          type: string
        reps:
          type: integer
        load:
          type: integer
      example:
        tille: random
        reps: 4
        load: 25
tags:
  - name: Workouts
    description: ""